refactor: consolidate base classes
