+++ src/galax/coordinates/_psp/pspt.py
@@ -1,13 +1,10 @@
 """galax: Galactic Dynamix in Jax."""
 
-__all__ = ["AbstractPhaseSpaceTimePosition", "PhaseSpaceTimePosition"]
+__all__ = ["PhaseSpaceTimePosition"]
 
-from dataclasses import replace
-from functools import partial
-from typing import TYPE_CHECKING, Any, NamedTuple, final
+from typing import Any, final
 
 import equinox as eqx
-import jax
 import jax.numpy as jnp
 from plum import convert
 
@@ -15,213 +12,11 @@ import quaxed.array_api as xp
 from coordinax import Abstract3DVector, Abstract3DVectorDifferential, Cartesian3DVector
 from unxt import Quantity
 
-from .base import AbstractPhaseSpacePositionBase
-from .utils import _p_converter, _q_converter, getitem_broadscalartime_index
-from galax.typing import BatchFloatQScalar, BatchVec7, BroadBatchFloatQScalar, QVec1
-from galax.units import UnitSystem, unitsystem
+from .base import AbstractPhaseSpaceTimePosition, ComponentShapeTuple
+from .utils import _p_converter, _q_converter
+from galax.typing import BatchVec7, BroadBatchFloatQScalar, QVec1
+from galax.units import unitsystem
 from galax.utils._shape import batched_shape, expand_batch_dims, vector_batched_shape
-
-if TYPE_CHECKING:
-    from typing import Self
-
-    from galax.potential._potential.base import AbstractPotentialBase
-
-
-class ComponentShapeTuple(NamedTuple):
-    """Component shape of the phase-space position."""
-
-    q: int
-    """Shape of the position."""
-
-    p: int
-    """Shape of the momentum."""
-
-    t: int
-    """Shape of the time."""
-
-
-class AbstractPhaseSpaceTimePosition(AbstractPhaseSpacePositionBase):
-    r"""Abstract base class of Phase-Space Positions with time.
-
-    The phase-space position is a point in the 7-dimensional phase space
-    :math:`\mathbb{R}^7` of a dynamical system. It is composed of the position
-    :math:`\boldsymbol{q}`, the conjugate momentum :math:`\boldsymbol{p}`, and
-    the time :math:`t`.
-
-    See Also
-    --------
-    :class:`~galax.coordinates.PhaseSpacePosition`
-        A phase-space position without time.
-    """
-
-    # TODO: shape Float[Array, "*#batch #time 3"]
-    q: eqx.AbstractVar[Abstract3DVector]
-    """Positions."""
-
-    # TODO: shape Float[Array, "*#batch #time 3"]
-    p: eqx.AbstractVar[Abstract3DVector]
-    """Conjugate momenta at positions ``q``."""
-
-    t: eqx.AbstractVar[BroadBatchFloatQScalar]
-    """Time corresponding to the positions and momenta."""
-
-    # ==========================================================================
-    # Array methods
-
-    def __getitem__(self, index: Any) -> "Self":
-        """Return a new object with the given slice applied."""
-        # Compute subindex
-        subindex = getitem_broadscalartime_index(index, self.t)
-        # Apply slice
-        return replace(self, q=self.q[index], p=self.p[index], t=self.t[subindex])
-
-    # ==========================================================================
-    # Convenience methods
-
-    def wt(self, *, units: UnitSystem) -> BatchVec7:
-        """Phase-space position as an Array[float, (*batch, 1+Q+P)].
-
-        This is the full phase-space position, including the time.
-
-        Parameters
-        ----------
-        units : `galax.units.UnitSystem`, keyword-only
-            The unit system If ``None``, use the current unit system.
-
-        Returns
-        -------
-        wt : Array[float, (*batch, 1+Q+P)]
-            The full phase-space position, including time on the first axis.
-
-        Examples
-        --------
-        We assume the following imports:
-
-        >>> from unxt import Quantity
-        >>> from coordinax import Cartesian3DVector, CartesianDifferential3D
-        >>> from galax.coordinates import PhaseSpaceTimePosition
-        >>> import galax.units as gu
-
-        We can create a phase-space position:
-
-        >>> psp = PhaseSpaceTimePosition(q=Quantity([1, 2, 3], "m"),
-        ...                              p=Quantity([4, 5, 6], "m/s"),
-        ...                              t=Quantity(7.0, "s"))
-        >>> psp.wt(units=gu.galactic)
-        Array([2.21816615e-13, 3.24077929e-20, 6.48155858e-20, 9.72233787e-20,
-               4.09084866e-06, 5.11356083e-06, 6.13627299e-06], dtype=float64)
-        """
-        batch, comps = self._shape_tuple
-        cart = self.represent_as(Cartesian3DVector)
-        q = xp.broadcast_to(convert(cart.q, Quantity), (*batch, comps.q))
-        p = xp.broadcast_to(convert(cart.p, Quantity), (*batch, comps.p))
-        t = xp.broadcast_to(self.t.decompose(units).value[..., None], (*batch, comps.t))
-        return xp.concat(
-            (t, q.decompose(units).value, p.decompose(units).value), axis=-1
-        )
-
-    # ==========================================================================
-    # Dynamical quantities
-
-    def potential_energy(
-        self, potential: "AbstractPotentialBase"
-    ) -> Quantity["specific energy"]:
-        r"""Return the specific potential energy.
-
-        .. math::
-
-            E_\Phi = \Phi(\boldsymbol{q})
-
-        Parameters
-        ----------
-        potential : `galax.potential.AbstractPotentialBase`
-            The potential object to compute the energy from.
-
-        Returns
-        -------
-        E : Array[float, (*batch,)]
-            The specific potential energy.
-
-        Examples
-        --------
-        We assume the following imports:
-
-        >>> from unxt import Quantity
-        >>> from coordinax import Cartesian3DVector, CartesianDifferential3D
-        >>> from galax.coordinates import PhaseSpaceTimePosition
-        >>> from galax.potential import MilkyWayPotential
-
-        We can construct a phase-space position:
-
-        >>> q = Cartesian3DVector(
-        ...     x=Quantity(1, "kpc"),
-        ...     y=Quantity([[1.0, 2, 3, 4], [1.0, 2, 3, 4]], "kpc"),
-        ...     z=Quantity(2, "kpc"))
-        >>> p = CartesianDifferential3D(
-        ...     d_x=Quantity(0, "km/s"),
-        ...     d_y=Quantity([[1.0, 2, 3, 4], [1.0, 2, 3, 4]], "km/s"),
-        ...     d_z=Quantity(0, "km/s"))
-        >>> w = PhaseSpaceTimePosition(q, p, t=Quantity(0, "Myr"))
-
-        We can compute the kinetic energy:
-
-        >>> pot = MilkyWayPotential()
-        >>> w.potential_energy(pot)
-        Quantity['specific energy'](Array(..., dtype=float64), unit='kpc2 / Myr2')
-        """
-        return potential.potential_energy(self.q, t=self.t)
-
-    @partial(jax.jit)
-    def energy(self, potential: "AbstractPotentialBase") -> BatchFloatQScalar:
-        r"""Return the specific total energy.
-
-        .. math::
-
-            E_K = \frac{1}{2} \\, |\boldsymbol{v}|^2
-            E_\Phi = \Phi(\boldsymbol{q})
-            E = E_K + E_\Phi
-
-        Parameters
-        ----------
-        potential : `galax.potential.AbstractPotentialBase`
-            The potential object to compute the energy from.
-
-        Returns
-        -------
-        E : Array[float, (*batch,)]
-            The kinetic energy.
-
-        Examples
-        --------
-        We assume the following imports:
-
-        >>> from unxt import Quantity
-        >>> from coordinax import Cartesian3DVector, CartesianDifferential3D
-        >>> from galax.coordinates import PhaseSpaceTimePosition
-        >>> from galax.potential import MilkyWayPotential
-
-        We can construct a phase-space position:
-
-        >>> q = Cartesian3DVector(
-        ...     x=Quantity(1, "kpc"),
-        ...     y=Quantity([[1.0, 2, 3, 4], [1.0, 2, 3, 4]], "kpc"),
-        ...     z=Quantity(2, "kpc"))
-        >>> p = CartesianDifferential3D(
-        ...     d_x=Quantity(0, "km/s"),
-        ...     d_y=Quantity([[1.0, 2, 3, 4], [1.0, 2, 3, 4]], "km/s"),
-        ...     d_z=Quantity(0, "km/s"))
-        >>> w = PhaseSpaceTimePosition(q, p, t=Quantity(0, "Myr"))
-
-        We can compute the kinetic energy:
-
-        >>> pot = MilkyWayPotential()
-        >>> w.energy(pot)
-        Quantity['specific energy'](Array(..., dtype=float64), unit='km2 / s2')
-        """
-        return self.kinetic_energy() + self.potential_energy(potential)
-
-
-###############################################################################
 
 
 @final
