+++ src/galax/dynamics/_dynamics/integrate/_api.py
@@ -2,7 +2,7 @@ __all__ = ["Integrator"]
 
 from typing import Any, Protocol, runtime_checkable
 
-from galax.coordinates import AbstractPhaseSpacePositionBase, PhaseSpaceTimePosition
+from galax.coordinates import AbstractPhaseSpaceTimePosition, PhaseSpaceTimePosition
 from galax.typing import FloatScalar, QVecTime, Vec6, VecTime
 from galax.units import UnitSystem
 from galax.utils.dataclasses import _DataclassInstance
@@ -45,7 +45,7 @@ class Integrator(_DataclassInstance, Protocol):
     def __call__(
         self,
         F: FCallable,
-        w0: AbstractPhaseSpacePositionBase | Vec6,
+        w0: AbstractPhaseSpaceTimePosition | Vec6,
         /,
         ts: QVecTime | VecTime,
         *,
@@ -58,7 +58,7 @@ class Integrator(_DataclassInstance, Protocol):
         F : FCallable, positional-only
             The function to integrate.
             (t, w, args) -> (v, a).
-        w0 : AbstractPhaseSpacePositionBase | Array[float, (6,)], positional-only
+        w0 : AbstractPhaseSpaceTimePosition | Array[float, (6,)], positional-only
             Initial conditions ``[q, p]``.
         ts : (Quantity | Array)[float, (T,)]
             Times to return the computation.
